// src/types/sanity.d.ts
export type SanitySectionNames = (typeof sections)[number]

export type SanityTourSectionNames = (typeof TourSections)[number]
// ...
export type SanityFAQSection = {
  _type: "faq_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  faqs?: SanityFAQ[]
}
// ...
export type SanityHeroCardSection = {
  _type: "hero_card_section"
  _id: string
  _key: string
  heading?: SanityLocaleString
  cta?: SanityLinkButton
  image?: SanityImage
}

export type SanityHeroSection = {
  _type: "hero_section"
  _id: string
  _key: string
  title?: SanityLocaleString
  subtitle?: SanityLocaleString
  image?: SanityImage
  cta_buttons?: SanityLinkButton[]
  cta_helper_text?: SanityLocaleString
  has_search?: boolean
  form_input_field?: SanityFormInputField
  scores?: SanityImage[]
}

export type SanityImageHeaderSection = {
  _type: "image_header_section"
  _id?: string
  _key?: string
  header?: SanityLocaleString
  image?: SanityImage | SanityPhoto
  content?: SanityLocaleText
}
// ...
export type SanityReviewsSection = {
  _type: "reviews_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  reviews?: {
    name?: SanityLocaleString
    time?: SanityLocaleString
    avatar?: SanityImage
    country?: SanityLocaleString
    rating?: number
    title?: SanityLocaleString
    text?: SanityLocaleString
  }[]
}
// ...
export type SanityFeaturedBlogsSection = {
  _type: "featured_blogs_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  featured_blogs?: ({
    _type: "reference"
    _id: string
    _ref: string
  } & SanityArticle)[]
}
// ...
export type SanityAtGlanceSection = {
  _type: "at_glance_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  facts?: {
    _type: "fact"
    _id: string
    _key: string
    title?: SanityLocaleString
    subtitle?: SanityLocaleString
    icon?: SanityIcon
  }[]
  useful_links_section?: {
    _type: "useful_links_section"
    _id: string
    _key: string
    title?: SanityLocaleString
    useful_links?: {
      _type: "useful_link"
      _id: string
      _key: string
      title?: SanityLocaleString
      url?: string
      icon?: SanityImage
    }[]
  }
}

export type SanityThingsToDo = {
  _type: "things_to_do"
  _id: string
  _key: string
  title?: SanityLocaleString
  description?: SanityLocaleText
  image?: SanityImage
  link?: SanityLink
}

export type SanityTopThingsSection = {
  _type: "top_things_section"
  _id: string
  _key: string
  destination?: {
    _type: "reference"
    _id: string
    _ref: string
  } & SanityDestinationPage
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  top_things?: ({ _type: "reference"; _ref: string } & SanityThingsToDo)[]
  cta?: SanityLocaleString
}

export type SanityTopToursSection = {
  _type: "top_tours_section"
  _id: string
  _key: string
  destination?: { _type: "reference"; _ref: string } & SanityDestinationPage
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  tours?: ({ _type: "reference"; _ref: string } & SanityTourPage)[]
}

export type SanityTourSelectionSection = {
  _type: "tour_selection_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  tags?: ({
    _type: "reference"
    _ref: string
  } & SanityTag)[]
  filters?: {
    _type: "filter"
    _id: string
    _key: string
    title?: SanityLocaleString
    filter_type?:
      | "destinations"
      | "special_offers"
      | "city_count"
      | "trip_type"
      | "price_range"
      | "duration"
  }[]
}
// ...
export type SanityFeaturedToursSection = {
  _type: "featured_tours_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  tour_cards?: {
    badge_content?: SanityLocaleString
    content?: {
      _type: "reference"
      _id: string
      _ref: string
    } & SanityTourPage
  }[]
}
// ...
export type SanityDestinationSection =
  | SanityImageHeaderSection
  | SanityAllBlogsSection
  | SanityReviewsSection
  | SanityFAQSection
  | SanityHeroCardSection
  | SanityFeaturedToursSection
  | SanityTourSelectionSection
  | SanityTopThingsSection
  | SanityAtGlanceSection
  | SanityContactAgentSection
  | SanityFeaturedBlogsSection

export type SanityDestinationSectionNames = SanityDestinationSection["_type"]

export type SanityContactAgentSection = {
  _type: "contact_agent_section"
  title?: SanityLocaleString
  cta?: SanityLinkButton
  hero_image?: SanityPhoto
}
// ...
export type SanityAllBlogsSection = {
  _type: "all_blogs_section"
  _id: string
  _key: string
  tagline?: SanityLocaleString
  title?: SanityLocaleString
  blogs?: SanityArticle[]
}